+++ UnitTests/HelperFunctions.h
@@ -10,5 +10,5 @@ namespace UnitTests
 	void LogEntropyGrid(Grid2D grid, WFCPosition expectedSize);
 	void LogDomainGrid(Grid2D grid, WFCPosition expectedSize); 
 	void createRules();
-	void AddToDomain(unsigned long tilesToAdd);
+	void AddToDomain(unsigned long long tilesToAdd);
 }
