+++ WaveFunctionCollapse/IWFCRule.cpp
@@ -1,18 +1,18 @@
 #include "IWFCRule.h"
 
-IWFCRule::IWFCRule(unsigned long goal, unsigned long tile, std::vector< WFCPosition*> localPositionsWeCareAbout)
+IWFCRule::IWFCRule(unsigned long long goal, unsigned long long tile, std::vector< WFCPosition*> localPositionsWeCareAbout)
 {
 	IWFCRule::goal = goal;
 	IWFCRule::tile = tile;
 	IWFCRule::localPositionsWeCareAbout = localPositionsWeCareAbout;
 }
 
-unsigned long const IWFCRule::GetGoal()
+unsigned long long const IWFCRule::GetGoal()
 {
 	return goal;
 }
 
-unsigned long const IWFCRule::GetTile()
+unsigned long long const IWFCRule::GetTile()
 {
 	return tile;
 }
